+++ src/ManagedNetworkFabric/ManagedNetworkFabric.Autorest/test/Invoke-AzNetworkFabricResyncNetworkBootstrapDevicePassword.Tests.ps1
@@ -1,0 +1,27 @@
+if(($null -eq $TestName) -or ($TestName -contains 'Invoke-AzNetworkFabricResyncNetworkBootstrapDevicePassword'))
+{
+  $loadEnvPath = Join-Path $PSScriptRoot 'loadEnv.ps1'
+  if (-Not (Test-Path -Path $loadEnvPath)) {
+      $loadEnvPath = Join-Path $PSScriptRoot '..\loadEnv.ps1'
+  }
+  . ($loadEnvPath)
+  $TestRecordingFile = Join-Path $PSScriptRoot 'Invoke-AzNetworkFabricResyncNetworkBootstrapDevicePassword.Recording.json'
+  $currentPath = $PSScriptRoot
+  while(-not $mockingPath) {
+      $mockingPath = Get-ChildItem -Path $currentPath -Recurse -Include 'HttpPipelineMocking.ps1' -File
+      $currentPath = Split-Path -Path $currentPath -Parent
+  }
+  . ($mockingPath | Select-Object -First 1).FullName
+}
+
+Describe 'Invoke-AzNetworkFabricResyncNetworkBootstrapDevicePassword' {
+    It 'Resync' {
+        {
+            Invoke-AzNetworkFabricResyncNetworkBootstrapDevicePassword -NetworkBootstrapDeviceName $global:config.networkBootstrapDevice.name -ResourceGroupName $global:config.common.resourceGroupName -SubscriptionId $global:config.common.subscriptionId
+        } | Should -Not -Throw
+    }
+
+    It 'ResyncViaIdentity' -skip {
+        { throw [System.NotImplementedException] } | Should -Not -Throw
+    }
+}