+++ src/ManagedNetworkFabric/ManagedNetworkFabric.Autorest/test/Update-AzNetworkFabricNetworkBootstrapInterface.Tests.ps1
@@ -1,0 +1,43 @@
+if(($null -eq $TestName) -or ($TestName -contains 'Update-AzNetworkFabricNetworkBootstrapInterface'))
+{
+  $loadEnvPath = Join-Path $PSScriptRoot 'loadEnv.ps1'
+  if (-Not (Test-Path -Path $loadEnvPath)) {
+      $loadEnvPath = Join-Path $PSScriptRoot '..\loadEnv.ps1'
+  }
+  . ($loadEnvPath)
+  $TestRecordingFile = Join-Path $PSScriptRoot 'Update-AzNetworkFabricNetworkBootstrapInterface.Recording.json'
+  $currentPath = $PSScriptRoot
+  while(-not $mockingPath) {
+      $mockingPath = Get-ChildItem -Path $currentPath -Recurse -Include 'HttpPipelineMocking.ps1' -File
+      $currentPath = Split-Path -Path $currentPath -Parent
+  }
+  . ($mockingPath | Select-Object -First 1).FullName
+}
+
+Describe 'Update-AzNetworkFabricNetworkBootstrapInterface' {
+    It 'UpdateExpanded' {
+        {
+            Update-AzNetworkFabricNetworkBootstrapInterface -NetworkBootstrapDeviceName $global:config.networkBootstrapDevice.name -Name $global:config.networkBootstrapInterface.name -ResourceGroupName $global:config.common.resourceGroupName -SubscriptionId $global:config.common.subscriptionId
+        } | Should -Not -Throw
+    }
+
+    It 'UpdateViaJsonString' -skip {
+        { throw [System.NotImplementedException] } | Should -Not -Throw
+    }
+
+    It 'UpdateViaJsonFilePath' -skip {
+        { throw [System.NotImplementedException] } | Should -Not -Throw
+    }
+
+    It 'UpdateViaIdentityNetworkBootstrapDeviceExpanded' -skip {
+        { throw [System.NotImplementedException] } | Should -Not -Throw
+    }
+
+    It 'UpdateViaIdentityNetworkBootstrapDevice' -skip {
+        { throw [System.NotImplementedException] } | Should -Not -Throw
+    }
+
+    It 'UpdateViaIdentityExpanded' -skip {
+        { throw [System.NotImplementedException] } | Should -Not -Throw
+    }
+}