+++ .azure-pipelines/PipelineSteps/analyze-steps.ps1
@@ -28,7 +28,8 @@ function Assert-ExitCodeSuccess {
 
 # Install PowerShell dependencies
 Write-Host -ForegroundColor Green "-------------------- Start installing PowerShell dependencies ... --------------------"
-Install-Module "platyPS", "PSScriptAnalyzer" -Repository PSGallery -Force -Confirm:$false -Scope CurrentUser
+Install-Module "Microsoft.PowerShell.PlatyPS" -MinimumVersion 1.0.2 -Repository PSGallery -Force -Confirm:$false -Scope CurrentUser
+Install-Module "PSScriptAnalyzer" -Repository PSGallery -Force -Confirm:$false -Scope CurrentUser
 Write-Host -ForegroundColor DarkGreen "-------------------- End installing PowerShell dependencies ... --------------------`n`n`n`n`n"
 
 # Install latest Az modules