+++ src/RecoveryServices/RecoveryServices/Models/PSObjects.cs
@@ -131,8 +131,14 @@ namespace Microsoft.Azure.Commands.RecoveryServices
                     this.Properties.PrivateEndpointConnections.Add(pec);
                 }                
             }
-            
-            if(vault.Properties.MonitoringSettings != null)
+
+            if(vault.Properties.CostManagementSettings != null)
+            {
+                this.Properties.CostManagementSettings = new CostManagementSettings();
+                this.Properties.CostManagementSettings.GranularityLevel = vault.Properties.CostManagementSettings.GranularityLevel;
+            }
+
+            if (vault.Properties.MonitoringSettings != null)
             {
                 this.Properties.AlertSettings = new AlertSettings();
 
@@ -284,6 +290,7 @@ namespace Microsoft.Azure.Commands.RecoveryServices
         public VaultPropertiesRedundancySettings RedundancySettings { get; set; }
         public SoftDeleteSettings SoftDeleteSettings {get; set; }
         public string MultiUserAuthorization { get; set; }
+        public CostManagementSettings CostManagementSettings { get; set; }
 
         public string SecureScore { get; set; }
         public string BcdrSecurityLevel { get; set; }
@@ -485,6 +492,27 @@ namespace Microsoft.Azure.Commands.RecoveryServices
         Disabled = 1,
         Unlocked,
         Locked
+    }
+
+    /// <summary>
+    /// Enum to define the Cost management granularity.
+    /// </summary>
+    public enum CostManagementGranularity
+    {
+        /// <summary>
+        /// Costs rolled up to vault resource (default)
+        /// </summary>
+        VaultLevel = 1,
+
+        /// <summary>
+        /// Costs shown per backup instance inside vault
+        /// </summary>
+        ProtectedItemLevel = 2,
+
+        /// <summary>
+        /// Costs shown per backup instance with parent resource tag
+        /// </summary>
+        ProtectedItemWithParentTag = 3
     }
 
     /// <summary>